requirements.txt
langchain
ipykernel
langchain_community
python-dotenv
pypdf
bs4
arxiv
pymupdf
wikipedia
lxml
langchain-openai
langchain-text-splitters
chromadb
sentence_transformers
langchain_huggingface
faiss-cpu
langchain_chroma
langchain-groq
fastapi
uvicorn
langserve
streamlit
langchain-core
